requirements.txt
OpenAI
python-dotenv
gradio